data_source/requirements.txt
pandas==2.2.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-decouple==3.8
